% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{celltype_specific_cosmx}
\alias{celltype_specific_cosmx}
\title{Extract cell type specific transcript count matrix and create Seurat object in cosmx data}
\usage{
celltype_specific_cosmx(
  sp_object,
  cell_types,
  conda.env.name = "STopover",
  fov_colname = "fov",
  cell_id_colname = "cell_ID",
  celltype_colname = "celltype",
  transcript_colname = "target",
  sc_norm_total = 1000,
  return_mode = "seurat"
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) containing CosMx SMI data to be used in calculating cell type-specific count matrix: non-normalized raw data should be in 'counts' slot}

\item{cell_types}{the cell types to extract cell type-specific count information}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{fov_colname}{column name for barcodes corresponding to fov (default = 'fov')}

\item{cell_id_colname}{column name for barcodes corresponding to cell ID (default = 'cell_ID')}

\item{celltype_colname}{column name for cell type annotation information saved in 'sp_object@assays$Spatial@misc' (default = 'celltype')}

\item{transcript_colname}{column name for the transcript name (default = 'target')}

\item{sc_norm_total}{scaling factor for the total count normalization per cell (default = 1e3)}

\item{return_mode}{what kind of data to return
\itemize{
\item 'seurat': return Seurat object that contains cell type-specific count matrix
\item 'anndata': return Anndata object that contains cell type-specific count matrix
}}
}
\value{
spatial data with cell type-specific count information in each grid
}
\description{
Extract cell type specific transcript count matrix and create Seurat object in cosmx data
}
