% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{convert_to_anndata}
\alias{convert_to_anndata}
\title{Convert Seurat object to anndata object}
\usage{
convert_to_anndata(
  sp_object,
  features = NULL,
  conda.env.name = "STopover",
  assay = "Spatial",
  slot = "data",
  add_coord = F,
  add_uns = list()
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used in converting}

\item{features}{a vector of features to convert to anndata: either gene symbols or column names for metadata (default = NULL)}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{assay}{assay to extract sparse matrix in Seurat object (default = 'Spatial')}

\item{slot}{slot to extract sparse matrix in Seurat object (default = 'data')}

\item{add_coord}{whether to add coordinate information in the .obs of anndata (default = F)}

\item{add_uns}{named list to add in the .uns of anndata (default = list())}
}
\value{
Anndata object that contains the count data of Seurat object (for the specific assay and slot) in .X and coordinates and metadata in .obs.
}
\description{
Convert spatial transcriptomic data from Seurat to anndata
}
