% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{install_load_env}
\alias{install_load_env}
\title{Install and load miniconda environment}
\usage{
install_load_env(conda.env.name = "STopover")
}
\arguments{
\item{conda.env.name}{name of the conda environment to use for STopover analysis (default: 'STopover')}
}
\description{
Install and load miniconda environment for implementation of STopover
}
