% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{preprocess_imageST}
\alias{preprocess_imageST}
\title{Load Image ST file as Seurat object (as a Slideseq class)}
\usage{
preprocess_imageST(
  sp_load_path = ".",
  sc_object = NULL,
  conda.env.name = "STopover",
  sc_celltype_colname = "celltype",
  ST_type = "cosmx",
  grid_method = "transcript",
  annot_method = "ingest",
  sc_norm_total = 1000,
  min_counts = 50,
  min_genes = 0,
  tx_file_name = "tx_file.csv",
  cell_exprmat_file_name = "exprMat_file.csv",
  cell_metadata_file_name = "metadata_file.csv",
  fov_colname = "fov",
  cell_id_colname = "cell_ID",
  tx_xcoord_colname = "x_global_px",
  tx_ycoord_colname = "y_global_px",
  transcript_colname = "target",
  meta_xcoord_colname = "CenterX_global_px",
  meta_ycoord_colname = "CenterY_global_px",
  x_bins = 100,
  y_bins = 100
)
}
\arguments{
\item{sp_load_path}{path to CosMx SMI data directory or .h5ad Anndata object (default = '.')}

\item{sc_object}{single-cell reference anndata/path to reference "*h5ad" file/Seurat object for cell type annotation of CosMx SMI data (default = NULL)
\itemize{
\item If NULL, then Leiden cluster numbers will be used to annotate CosMx SMI data
}}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{sc_celltype_colname}{column name for cell type annotation information in metadata of single-cell (default = 'celltype')}

\item{sc_norm_total}{scaling factor for the total count normalization per cell (default = 1e3)}

\item{tx_file_name}{CosMx file for transcript count (default = 'tx_file.csv')}

\item{cell_exprmat_file_name}{CosMx file for cell-level expression matrix (default = 'exprMat_file.csv')}

\item{cell_metadata_file_name}{CosMx file for cell-level metadata (default = 'metadata_file.csv')}

\item{fov_colname}{column name for barcodes corresponding to fov (default = 'fov')}

\item{cell_id_colname}{column name for barcodes corresponding to cell ID (default = 'cell_ID')}

\item{tx_xcoord_colname}{column name for global x coordinates of the transcript (default = 'x_global_px')}

\item{tx_ycoord_colname}{column name for global y coordinates of the transcript (default = 'y_global_px')}

\item{transcript_colname}{column name for the transcript name (default = 'target')}

\item{meta_xcoord_colname}{column name for global x, y coordinates in cell-level metadata file (default = 'CenterX_global_px')}

\item{meta_ycoord_colname}{column name for global x, y coordinates in cell-level metadata file (default = 'CenterY_global_px')}

\item{x_bins}{number of bins to divide the image-based ST data (for grid-based aggregation) (default = 100)}

\item{y_bins}{number of bins to divide the image-based ST data (for grid-based aggregation) (default = 100)}

\item{ST_type:}{type of the ST data to be read: cosmx, xenium, merfish (default: 'cosmx')}

\item{grid_method:}{type of the method to assign transcript to grid, either transcript coordinate based method and cell coordinate based method (default='transcript')}

\item{annot_method:}{cell type annotation method to use. Either 'ingest' or 'tacco' (default='ingest')}

\item{min_counts:}{minimum number of counts required for a cell to pass filtering (scanpy.pp.filter_cells) (default = 50)}

\item{min_genes:}{minimum number of genes expressed required for a cell to pass filtering (scanpy.pp.filter_cells) (default = 0)}
}
\value{
list of Seurat objects for grid-based and cell-level data, sequentially. The objects contains grid-based or cell-level log-normalized count matrix in 'counts & data' slots, coordinates in '@images$image', and transcript information in '@assays$Spatial@misc' (only for grid-based data)
}
\description{
Convert Seurat object for spatial transcriptomic data to anndata
}
