% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{return_celltalkdb}
\alias{return_celltalkdb}
\title{Extract CellTalkDB as dataframe}
\usage{
return_celltalkdb(lr_db_species = "human", conda.env.name = "STopover")
}
\arguments{
\item{lr_db_species}{select species to utilize in CellTalkDB database (default = 'human')}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}
}
\description{
Extract CellTalkDB as dataframe
}
