% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_R.R
\name{return_lr_db}
\alias{return_lr_db}
\title{Extract ligand-receptor database as dataframe}
\usage{
return_lr_db(
  lr_db_species = "human",
  db_name = "CellTalk",
  conda.env.name = "STopover"
)
}
\arguments{
\item{lr_db_species}{select species to utilize in CellTalkDB database (default = 'human')}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{db_name:}{name of the ligand-receptor database to use: either 'CellTalk', 'CellChat', or 'Omnipath' (default = 'CellTalk')}
}
\description{
Extract ligand-receptor database as dataframe
}
