% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_sim_R.R
\name{topological_similarity}
\alias{topological_similarity}
\title{Calculate topological overlap of feature pairs and return Jaccard dataframe}
\usage{
topological_similarity(
  sp_object,
  feat_pairs = data.frame(),
  use_lr_db = F,
  lr_db_species = "human",
  min_size = 20,
  fwhm = 2.5,
  thres_per = 30,
  conda.env.name = "STopover",
  assay = "Spatial",
  slot = "data",
  lognorm = F,
  J_result_name = "result",
  num_workers = NULL
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used in calculating topological similarity
\itemize{
\item select which slot to use and whether to perform log-normalization before STopover analysis
\item if log-normalized data is included in 'data' slot, then set slot as 'data' and 'lognorm' as F, which are default values
}}

\item{feat_pairs}{list of features with the format \link{('A','B'),('C','D')} or the pandas equivalent (default = data.frame())
\itemize{
\item (A and C) should to be same data format: all in metadata (.obs.columns) or all in gene names (.var.index).
\item (C and D) should be same data format: all in metadata (.obs.columns) or all in gene names (.var.index).
\item If the data format is not same the majority of the data format will be automatically searched and the rest of the features with different format will be removed from the pairs (default = '.').
}}

\item{use_lr_db}{whether to use list of features in CellTalkDB L-R database (default = F)}

\item{lr_db_species}{select species to utilize in CellTalkDB database (default = 'human')}

\item{min_size}{minimum size of a connected component: number of spots/grids (default = 20)}

\item{fwhm}{full width half maximum value for the Gaussian smoothing kernel as the multiple of the central distance between the adjacent spots (default = 2.5)}

\item{thres_per}{lower percentile value threshold to remove the connected components (default = 30)}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{assay}{assay to extract sparse matrix in Seurat object (default = 'Spatial')}

\item{slot}{slot to extract sparse matrix in Seurat object (default = 'data': expecting lognormalized matrix in 'data')}

\item{lognorm}{whether to lognormalize the sparse matrix in the slot before calculating topological similarity (default = F)}

\item{J_result_name}{the name of the jaccard index data file name (default = 'result')}

\item{num_workers}{number of the workers for the multiprocessing (default = NULL)
\itemize{
\item if NULL, the maximum number of CPUs will be used for the multiprocessing
}}
}
\value{
spatial data (Seurat object) with connected component location described as integers in metadata (meta.data)
}
\description{
Calculate Jaccard index between topological connected components of feature pairs and return connected component location in meta.data and dataframe with Jaccard indexes in misc
}
