% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_sim_R.R
\name{topological_similarity_celltype_pair}
\alias{topological_similarity_celltype_pair}
\title{Calculate spatial overlap of feature pairs from a specific cell type pair and return Jaccard dataframe}
\usage{
topological_similarity_celltype_pair(
  sp_object,
  celltype_x = "",
  celltype_y = "",
  feat_pairs = data.frame(),
  conda.env.name = "STopover",
  fov_colname = "fov",
  cell_id_colname = "cell_ID",
  celltype_colname = "celltype",
  transcript_colname = "target",
  sc_norm_total = 1000,
  assay = "Spatial",
  slot = "data",
  lognorm = F,
  min_size = 20,
  fwhm = 2.5,
  thres_per = 30,
  use_lr_db = F,
  lr_db_species = "human",
  J_result_name = "result"
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used in calculating topological similarity
\itemize{
\item select which slot to use and whether to perform log-normalization before STopover analysis
\item if log-normalized data is included in 'data' slot, then set slot as 'data' and 'lognorm' as F, which are default values
}}

\item{celltype_x}{cell type corresponding to the first feature (default = '')}

\item{celltype_y}{cell type corresponding to the second feature (default = '')}

\item{feat_pairs}{list of features with the format \link{('A','B'),('C','D')} or the pandas equivalent (default = data.frame())
\itemize{
\item (A and C) should to be same data format: all in metadata (.obs.columns) or all in gene names (.var.index).
\item (C and D) should be same data format: all in metadata (.obs.columns) or all in gene names (.var.index).
\item If the data format is not same the majority of the data format will be automatically searched and the rest of the features with different format will be removed from the pairs (default = '.').
}}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{fov_colname}{column name for barcodes corresponding to fov (default = 'fov')}

\item{cell_id_colname}{column name for barcodes corresponding to cell ID (default = 'cell_ID')}

\item{celltype_colname}{column name for cell type annotation information saved in 'sp_object@assays$Spatial@misc' (default = 'celltype')}

\item{transcript_colname}{column name for the transcript name (default = 'target')}

\item{sc_norm_total}{scaling factor for the total count normalization per cell (default = 1e3)}

\item{assay}{assay to extract sparse matrix in Seurat object while calculating topological similarity between cell types in visium dataset (default = 'Spatial')}

\item{slot}{slot to extract sparse matrix in Seurat object while calculating topological similarity between cell types in visium dataset (default = 'data')}

\item{lognorm}{whether to lognormalize the sparse matrix in the slot before calculating topological similarity in visium dataset (default = F)}

\item{min_size}{minimum size of a connected component: number of spots/grids (default = 20)}

\item{fwhm}{full width half maximum value for the Gaussian smoothing kernel as the multiple of the central distance between the adjacent spots (default = 2.5)}

\item{thres_per}{lower percentile value threshold to remove the connected components (default = 30)}

\item{use_lr_db}{whether to use list of features in CellTalkDB L-R database (default = F)}

\item{lr_db_species}{select species to utilize in CellTalkDB database (default = 'human')}

\item{J_result_name}{the name of the jaccard index data file name (default = 'result')}
}
\value{
spatial data (Seurat object) with connected component location described as integers in metadata (meta.data)
}
\description{
Calculate spatial overlap between topological connected components of feature pairs from a specific cell type pair and return connected component location in meta.data and dataframe with Jaccard indexes in misc
}
