% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_vis_R.R
\name{vis_all_connected}
\alias{vis_all_connected}
\title{Visualizing all connected components of feature x and y}
\usage{
vis_all_connected(
  sp_object,
  feat_name_x = "",
  feat_name_y = "",
  celltype_x = NULL,
  celltype_y = NULL,
  conda.env.name = "STopover",
  dot_size = 1.8,
  alpha_img = 0.8,
  alpha = 0.8,
  vis_jaccard = T,
  subset_by_slide = F,
  slide_names = names(sp_object@images),
  slide_titles = NULL,
  slide_ncol = 2,
  title_fontsize = 15,
  legend_fontsize = 10,
  add_title_text = "",
  crop_image = T,
  return_plot = F,
  save = F,
  save_path = ".",
  save_name_add = "",
  dpi = 100,
  fig_width = 4,
  fig_height = 4
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used}

\item{feat_name_x}{name of the feature x (default = '')}

\item{feat_name_y}{name of the feature y (default = '')}

\item{celltype_x}{cell type corresponding to the first feature if the cell type specific data is provided (default = NULL)}

\item{celltype_y}{cell type corresponding to the second feature if the cell type specific data is provided (default = NULL)}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{dot_size}{size of the spot/grid visualized on the tissue (default =  1.8)}

\item{alpha_img}{transparency of the background image (not available for cosmx data) (default = 0.8)}

\item{alpha}{transparency of the colored spot (default = 0.8)}

\item{vis_jaccard}{whether to visualize jaccard index on right corner of the plot (default = T)}

\item{subset_by_slide}{whether to select the certain slides for the visualization (if there are multiple slides in sp_object) (default = F)}

\item{slide_names}{name of the slides to select (among 'names(sp_object@images)') (default = names(sp_object@images))}

\item{slide_titles}{title of the slides to visualize in the plot (default = NULL)}

\item{slide_ncol}{number of images to visualize in the column (default = 2)}

\item{title_fontsize}{fontsize of the figure title (default = 15)}

\item{legend_fontsize}{fontsize of the figure legend (default = 10)}

\item{add_title_text}{the text to add in the figure title (default = ')}

\item{crop_image}{whether to crop the image (default = T)}

\item{return_plot}{whether to return the plot as list (default = F)}

\item{save}{whether to save the image (default = F)}

\item{save_path}{path to save the image (default = '.')}

\item{save_name_add}{the text to add in the file name (default = ')}

\item{dpi}{dpi to save the image (default = 100)}

\item{fig_width}{figure width in ggsave (default = 5)}

\item{fig_height}{figure height in ggsave (default = 5)}
}
\value{
list of plots
}
\description{
Visualizing connected component locations of feature x and feature y on the tissue and return plots or save plots if designated
}
