% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_vis_R.R
\name{vis_diff_inc_lr_pairs}
\alias{vis_diff_inc_lr_pairs}
\title{Visualizing upregulated LR interactions and their corresponding GO terms using a heatmap}
\usage{
vis_diff_inc_lr_pairs(
  sp_object,
  ref_group,
  comp_group,
  logFC_cutoff = 1,
  J_comp_cutoff = 0.2,
  go_species = c("human", "mouse"),
  db_name = "CellTalk",
  ontology_cat = c("BP", "CC", "MF", "all"),
  padjust_method = "BH",
  padjust_cutoff = 0.05,
  top_n = 10,
  heatmap_max = 10,
  title_fontsize = 12,
  xaxis_title_fontsize = 10,
  yaxis_title_fontsize = 10,
  angle_col = 45,
  colorbar_palette = rownames(RColorBrewer::brewer.pal.info),
  legend_loc = "right",
  legend_fontsize = 10,
  save_plot = F,
  save_path = ".",
  save_name = "heatmap_GO_LR_int",
  fig_width = 21,
  fig_height = 10,
  dpi = 150,
  return_results = T
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used}

\item{ref_group}{vector containing slide name(s) corresponding to the group that are considered as a baseline (should be among names(sp_object@images))}

\item{comp_group}{vector containing slide name(s) corresponding to the group in which upregulated LR paris are tested (should be among names(sp_object@images))}

\item{logFC_cutoff}{log fold change cutoff of J_comp for the differentially upregulated LR pairs in 'comp_group' compared to 'ref_group' (default = 1)}

\item{J_comp_cutoff}{J_comp cutoff in the comp_group (default = 0.2)}

\item{go_species}{species of the given dataset (default = "human")}

\item{ontology_cat}{category of GO terms to show in the heatmap (default ="BP")}

\item{padjust_method}{method for adjusting p-values during overrepresentation tests to identify significant GO terms (default ="BH")}

\item{padjust_cutoff}{adjusted p-value cutoff to select the GO terms (default = 0.05)}

\item{top_n}{the top GO terms to show in the heatmap (default=10)}

\item{heatmap_max}{the maximum value to show in the colorbar (default=10)}

\item{title_fontsize}{the title fontsize in the heatmap (default = 12)}

\item{xaxis_title_fontsize}{the x axis label size in the heatmap (default = 12)}

\item{yaxis_title_fontsize}{the y axis label size in the heatmap (default = 12)}

\item{angle_col}{the angle of the column label in the heatmap (refer to the pheatmap 'angle_col') (default = 45)}

\item{colorbar_palette}{the vector of colors to be used in the colormap (select among the grDevices::hcl.pals())}

\item{legend_loc}{location of the figure legend (default = 'right')}

\item{save_plot}{whether to save the heatmap (default = F)}

\item{save_path}{the path to save the heatmap (default = '.')}

\item{save_name}{the name of the heatmap to save (default = 'heatmap_GO_LR_int.png')}

\item{fig_width}{the width of the heatmap to save (default = 21)}

\item{fig_height}{the height of the heatmap to save (default = 10)}

\item{dpi}{the dpi of the heatmap to save (default = 150)}

\item{return_results}{return a dataframe showing the list of top upregulated LR interaction and a heatmap}

\item{db_name:}{name of the ligand-receptor database to use: either 'CellTalk', 'CellChat', or 'Omnipath' (default = 'CellTalk')}
}
\value{
a list of dataframe showing the list of top differentially increased LR interaction in comp_group compared to ref_group and a heatmap
}
\description{
Spatial LR interaction profiles are estimated by STopover and differentially upregulated LR interactions in 'comp_group' compared to 'ref_group' are selected (threshold: Jcomp>J_comp_cutoff & fold change of Jcomp>J_comp_cutoff). The enrichment analysis is performed for the genes in LR pairs and the top 10 enriched GO terms with the lowest adjusted p-values were selected (threshold: p.adjust<padjust_cutoff). LR pairs in which either ligand or receptor gene intersected with the genes composing the top GO terms were listed. The upregulated functional terms, the intersecting LR pairs, and fold change of Jcomp are visualized in a heatmap.
}
