% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_vis_R.R
\name{vis_jaccard_top_n_pair}
\alias{vis_jaccard_top_n_pair}
\title{Visualizing top connected components with the highest local Jaccard index}
\usage{
vis_jaccard_top_n_pair(
  sp_object,
  feat_name_x = "",
  feat_name_y = "",
  celltype_x = NULL,
  celltype_y = NULL,
  top_n = 2,
  conda.env.name = "STopover",
  dot_size = 1.8,
  alpha_img = 0.8,
  alpha = 0.8,
  vis_jaccard = T,
  slide_name = names(sp_object@images)[1],
  slide_title = NULL,
  slide_ncol = 2,
  title_fontsize = 15,
  legend_loc = "right",
  legend_fontsize = 10,
  add_title_text = "",
  crop_image = F,
  return_plot = F,
  return_sp_object = F,
  save = F,
  save_path = ".",
  save_name_add = "",
  dpi = 100,
  fig_width = 5,
  fig_height = 5
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used}

\item{feat_name_x}{name of the feature x (default = '')}

\item{feat_name_y}{name of the feature y (default = '')}

\item{celltype_x}{cell type corresponding to the first feature if the cell type specific data is provided (default = NULL)}

\item{celltype_y}{cell type corresponding to the second feature if the cell type specific data is provided (default = NULL)}

\item{top_n}{the number of the top connected component pairs with the highest Jaccard similarity index}

\item{conda.env.name}{name of the conda environment to use for STopover analysis (default = 'STopover')}

\item{dot_size}{size of the spot/grid visualized on the tissue (default =  1.8)}

\item{alpha_img}{transparency of the background image (not available for imageST data) (default = 0.8)}

\item{alpha}{transparency of the colored spot (default = 0.8)}

\item{vis_jaccard}{whether to visualize jaccard index on right corner of the plot (default = T)}

\item{slide_name}{name of one slide to select (among 'names(sp_object@images)') (default = names(sp_object@images))}

\item{slide_title}{title of the selected slide to visualize in the plot (default = NULL)}

\item{slide_ncol}{number of images to visualize in the column (default = 2)}

\item{title_fontsize}{fontsize of the figure title (default = 15)}

\item{legend_loc}{location of the figure legend (default = 'right')}

\item{legend_fontsize}{fontsize of the figure legend (default = 10)}

\item{add_title_text}{the text to add in the figure title (default = '')}

\item{crop_image}{whether to crop the image (default = F)}

\item{return_plot}{whether to return the plot as list (default = F)}

\item{return_sp_object}{whether to return the sp_object containing top n connected component location saved in metadata (default = F)}

\item{save}{whether to save the image (default = F)}

\item{save_path}{path to save the image (default = '.')}

\item{save_name_add}{the text to add in the file name (default = '')}

\item{dpi}{dpi to save the image (default = 100)}

\item{fig_width}{figure width in ggsave (default = 5)}

\item{fig_height}{figure height in ggsave (default = 5)}
}
\value{
list of plots or Seurat object containing top n connected component location
}
\description{
Visualizing top connected component locations of feature x and feature y having the highest local Jaccard index and return plots or save plots if designated
}
