% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_vis_R.R
\name{vis_spatial_cosmx}
\alias{vis_spatial_cosmx}
\title{Visualizing CosMx SMI data on the grid}
\usage{
vis_spatial_cosmx(
  sp_object,
  feature,
  plot_title = feature,
  title_fontsize = 12,
  color_dis = c("#A2E1CA", "#FBBC05", "#4285F4", "#34A853"),
  color_cont = "RdPu",
  vmin = NULL,
  vmax = NULL,
  legend_loc = "right",
  legend_fontsize = 12
)
}
\arguments{
\item{sp_object}{spatial data (Seurat object) to be used}

\item{feature}{feature to visualize on the grid}

\item{plot_title}{title of the plot (default = feature)}

\item{title_fontsize}{fontsize of the plot title (default = 12)}

\item{color_dis}{if the feature is discrete value, provide vector of colors to distinguish the values (default = c("#A2E1CA","#FBBC05","#4285F4","#34A853"))}

\item{color_cont}{if the feature is a continuous variable, provide color palette to use from RColorBrewer::brewer.pal.info (default = "RdPu")}

\item{vmin}{if the feature is a continuous variable, provide the minimum of the colorbar (default = NULL)}

\item{vmax}{if the feature is a continuous variable, provide the maximum of the colorbar (default = NULL)}

\item{legend_loc}{location of the legend (default= 'right')}

\item{legend_fontsize}{fontsize of the legend title (default = 12)}
}
\value{
ggplot object
}
\description{
Visualizing CosMx SMI data with ggplot2
}
